+++ skhubness/neighbors/lof.py
@@ -48,7 +48,7 @@ class LocalOutlierFactor(NeighborsBase, KNeighborsMixin, UnsupervisedMixin,
         - 'hnsw' will use :class:`LegacyHNSW`
         - 'lsh' will use :class:`PuffinnLSH`
         - 'falconn_lsh' will use :class:`FalconnLSH`
-        - 'nng' will use :class:`NNG`
+        - 'nng' will use :class:`LegacyNNG`
         - 'rptree' will use :class:`LegacyRandomProjectionTree`
         - 'ball_tree' will use :class:`BallTree`
         - 'kd_tree' will use :class:`KDTree`