+++ skhubness/neighbors/tests/test_onng.py
@@ -6,7 +6,7 @@ from sklearn.datasets import make_classification
 from sklearn.utils.estimator_checks import check_estimator
 from sklearn.utils.testing import assert_array_equal, assert_array_almost_equal
 from sklearn.utils.testing import assert_raises
-from skhubness.neighbors import NNG, NearestNeighbors
+from skhubness.neighbors import LegacyNNG, NearestNeighbors
 
 
 @pytest.mark.skipif(sys.platform == 'win32', reason='NGT not supported on Windows.')
@@ -22,8 +22,8 @@ def test_return_correct_number_of_neighbors(n_candidates: int,
                                             verbose: bool):
     n_samples = 100
     X, y = make_classification(n_samples=n_samples)
-    ann = NNG(n_candidates=n_candidates, verbose=verbose)\
-        if set_in_constructor else NNG(verbose=verbose)
+    ann = LegacyNNG(n_candidates=n_candidates, verbose=verbose)\
+        if set_in_constructor else LegacyNNG(verbose=verbose)
     ann.fit(X, y)
     X_query = None if search_among_indexed else X
     neigh = ann.kneighbors(X_query, return_distance=return_distance) if set_in_constructor\
@@ -44,22 +44,22 @@ def test_return_correct_number_of_neighbors(n_candidates: int,
 @pytest.mark.parametrize('metric', ['invalid', None])
 def test_invalid_metric(metric):
     X, y = make_classification(n_samples=10, n_features=10)
-    ann = NNG(metric=metric)
+    ann = LegacyNNG(metric=metric)
     with assert_raises(ValueError):
         _ = ann.fit(X, y)
 
 
 @pytest.mark.skipif(sys.platform == 'win32', reason='NGT not supported on Windows.')
-@pytest.mark.parametrize('metric', NNG.valid_metrics)
+@pytest.mark.parametrize('metric', LegacyNNG.valid_metrics)
 @pytest.mark.parametrize('n_jobs', [-1, 1, None])
 @pytest.mark.parametrize('verbose', [0, 1])
 def test_kneighbors_with_or_without_distances(metric, n_jobs, verbose):
     n_samples = 100
     X = np.random.RandomState(1235232).rand(n_samples, 2)
-    ann = NNG(metric=metric,
-              n_jobs=n_jobs,
-              verbose=verbose,
-              )
+    ann = LegacyNNG(metric=metric,
+                    n_jobs=n_jobs,
+                    verbose=verbose,
+                    )
     ann.fit(X)
     neigh_dist_self, neigh_ind_self = ann.kneighbors(X, return_distance=True)
     ind_only_self = ann.kneighbors(X, return_distance=False)
@@ -83,13 +83,13 @@ def test_kneighbors_with_or_without_distances(metric, n_jobs, verbose):
 
 
 @pytest.mark.skipif(sys.platform == 'win32', reason='NGT not supported on Windows.')
-@pytest.mark.parametrize('metric', NNG.valid_metrics)
+@pytest.mark.parametrize('metric', LegacyNNG.valid_metrics)
 def test_kneighbors_with_or_without_self_hit(metric):
     X = np.random.RandomState(1245544).rand(50, 2)
     n_candidates = 5
-    ann = NNG(metric=metric,
-              n_candidates=n_candidates,
-              )
+    ann = LegacyNNG(metric=metric,
+                    n_candidates=n_candidates,
+                    )
     ann.fit(X)
     ind_self = ann.kneighbors(X, n_candidates=n_candidates+1, return_distance=False)
     ind_no_self = ann.kneighbors(n_candidates=n_candidates, return_distance=False)
@@ -105,11 +105,11 @@ def test_kneighbors_with_or_without_self_hit(metric):
 @pytest.mark.skipif(sys.platform == 'win32', reason='NGT not supported on Windows.')
 def test_squared_euclidean_same_neighbors_as_euclidean():
     X, y = make_classification()
-    ann = NNG(metric='euclidean')
+    ann = LegacyNNG(metric='euclidean')
     ann.fit(X, y)
     neigh_dist_eucl, neigh_ind_eucl = ann.kneighbors(X)
 
-    ann = NNG(metric='sqeuclidean')
+    ann = LegacyNNG(metric='sqeuclidean')
     ann.fit(X, y)
     neigh_dist_sqeucl, neigh_ind_sqeucl = ann.kneighbors(X)
 
@@ -124,7 +124,7 @@ def test_same_neighbors_as_with_exact_nn_search():
     nn = NearestNeighbors()
     nn_dist, nn_neigh = nn.fit(X).kneighbors(return_distance=True)
 
-    ann = NNG()
+    ann = LegacyNNG()
     ann_dist, ann_neigh = ann.fit(X).kneighbors(return_distance=True)
 
     assert_array_almost_equal(ann_dist, nn_dist, decimal=5)
@@ -133,13 +133,13 @@ def test_same_neighbors_as_with_exact_nn_search():
 
 @pytest.mark.skipif(sys.platform == 'win32', reason='NGT not supported on Windows.')
 def test_is_valid_estimator_in_persistent_memory():
-    check_estimator(NNG)
+    check_estimator(LegacyNNG)
 
 
 @pytest.mark.skipif(sys.platform == 'win32', reason='NGT not supported on Windows.')
 @pytest.mark.xfail(reason='ngtpy.Index can not be pickled as of v1.7.6')
 def test_is_valid_estimator_in_main_memory():
-    check_estimator(NNG(index_dir=None))
+    check_estimator(LegacyNNG(index_dir=None))
 
 
 @pytest.mark.skipif(sys.platform == 'win32', reason='NGT not supported on Windows.')
@@ -149,7 +149,7 @@ def test_memory_mapped(index_dir):
                                n_features=5,
                                random_state=123,
                                )
-    ann = NNG(index_dir=index_dir)
+    ann = LegacyNNG(index_dir=index_dir)
     if isinstance(index_dir, str) or index_dir is None:
         ann.fit(X, y)
         _ = ann.kneighbors(X)
@@ -166,13 +166,13 @@ def test_nng_optimization():
                                n_redundant=0,
                                random_state=123,
                                )
-    ann = NNG(index_dir='/dev/shm',
-              optimize=True,
-              edge_size_for_search=40,
-              edge_size_for_creation=10,
-              epsilon=0.1,
-              n_jobs=2,
-              )
+    ann = LegacyNNG(index_dir='/dev/shm',
+                    optimize=True,
+                    edge_size_for_search=40,
+                    edge_size_for_creation=10,
+                    epsilon=0.1,
+                    n_jobs=2,
+                    )
     ann.fit(X, y)
     _ = ann.kneighbors(X, )
     _ = ann.kneighbors()